feat: Sprint 2+3 SSR migration, SEO metadata, analytics & a11y improvements
SSR conversions (Tasks 25-32):
- Convert matches/[id], matches/[id]/[incidentSlug], hakemler/[slug],
  superlig-tartismali-pozisyonlar to server components with Prisma queries
- Each page has a Client component wrapper for interactive state

SEO (Tasks 49-57, 60-62):
- generateMetadata on all public pages with canonical URLs and OG data
- JSON-LD SportsEvent (match layout), Person (referee layout),
  WebSite + Organization (root layout)

Analytics & Monitoring (Tasks 69-70):
- GA4 integration via NEXT_PUBLIC_GA_ID env var
- Sentry @sentry/nextjs setup with client/server/edge configs

Cron Jobs (Task 93):
- vercel.json cron schedule (every 6h) + /api/cron/scrape endpoint

Performance (Tasks 35-38 already done, Task 88):
- Skeleton components: MatchCardSkeleton, IncidentCardSkeleton, RefereeCardSkeleton

UI/UX (Task 89):
- sonner Toaster added to root layout for global toast feedback

Accessibility (Tasks 117-118):
- aria-label on icon-only report buttons in CommentSection
- Focus trap implementation in AuthModal (keyboard nav + Escape close)
