+++ .env.example
@@ -8,6 +8,14 @@ AUTH_SECRET="openssl rand -base64 32 ile üretin"
 AUTH_GOOGLE_ID=""
 AUTH_GOOGLE_SECRET=""
 
+# Analytics (opsiyonel)
+NEXT_PUBLIC_GA_ID=""
+NEXT_PUBLIC_CLARITY_ID=""
+SENTRY_DSN=""
+
 # Diğer
 OPENAI_API_KEY=""
 ADMIN_SECRET=""
+CRON_SECRET=""
+SENTRY_ORG=""
+SENTRY_PROJECT=""
