+++ sentry.client.config.ts
@@ -1,0 +1,10 @@
+import * as Sentry from "@sentry/nextjs";
+
+Sentry.init({
+  dsn: process.env.SENTRY_DSN,
+  enabled: !!process.env.SENTRY_DSN && process.env.NODE_ENV === "production",
+  tracesSampleRate: 0.1,
+  replaysOnErrorSampleRate: 1.0,
+  replaysSessionSampleRate: 0.05,
+  integrations: [Sentry.replayIntegration()],
+});
