+++ src/app/matches/[id]/[incidentSlug]/page.tsx
@@ -1,132 +1,170 @@
-"use client";
+import type { Metadata } from "next";
+import { prisma } from "@/database/db";
+import { buildMatchSlug, getShortIdFromIncidentSlug } from "@/lib/slug";
+import { getResolvedVideoUrl } from "@/lib/incident-api";
+import { notFound } from "next/navigation";
+import IncidentPageClient from "./IncidentPageClient";
 
-import { useState, useEffect, useCallback, use } from "react";
-import { useRouter } from "next/navigation";
-import CommentSection from "@/components/CommentSection";
-import IncidentDetailTemplate, { type IncidentDetailData } from "@/components/IncidentDetailTemplate";
-import { matchUrl } from "@/lib/links";
-import { Loader2 } from "lucide-react";
+export const revalidate = 60;
 
-const TYPE_LABELS: Record<string, { label: string }> = {
-  POSSIBLE_PENALTY: { label: "Penaltı Pozisyonu" },
-  PENALTY: { label: "Penaltı Kararı" },
-  POSSIBLE_OFFSIDE_GOAL: { label: "Ofsayt Tartışması" },
-  OFFSIDE: { label: "Ofsayt Kararı" },
-  MISSED_RED_CARD: { label: "Verilmeyen Kırmızı Kart" },
-  MISSED_YELLOW: { label: "Verilmeyen Sarı Kart" },
-  RED_CARD: { label: "Kırmızı Kart" },
-  YELLOW_CARD: { label: "Sarı Kart" },
-  VAR_CONTROVERSY: { label: "VAR Tartışması" },
-  GOAL_DISALLOWED: { label: "İptal Edilen Gol" },
-  FOUL: { label: "Faul Kararı" },
-  HANDBALL: { label: "El ile Temas" },
+const SITE_URL = "https://varodasi.com";
+
+const TYPE_LABELS: Record<string, string> = {
+  POSSIBLE_PENALTY: "Penaltı Pozisyonu",
+  PENALTY: "Penaltı Kararı",
+  POSSIBLE_OFFSIDE_GOAL: "Ofsayt Tartışması",
+  OFFSIDE: "Ofsayt Kararı",
+  MISSED_RED_CARD: "Verilmeyen Kırmızı Kart",
+  MISSED_YELLOW: "Verilmeyen Sarı Kart",
+  RED_CARD: "Kırmızı Kart",
+  YELLOW_CARD: "Sarı Kart",
+  VAR_CONTROVERSY: "VAR Tartışması",
+  GOAL_DISALLOWED: "İptal Edilen Gol",
+  FOUL: "Faul Kararı",
+  HANDBALL: "El ile Temas",
 };
 
-function parseJson<T>(raw: unknown, fallback: T): T {
-  if (Array.isArray(raw)) return raw as T;
-  if (typeof raw === "string") {
-    try {
-      return JSON.parse(raw) as T;
-    } catch {
-      return fallback;
-    }
-  }
-  return fallback;
-}
+export async function generateMetadata({
+  params,
+}: {
+  params: Promise<{ id: string; incidentSlug: string }>;
+}): Promise<Metadata> {
+  const { id: matchSlug, incidentSlug } = await params;
 
-function toTemplateData(incident: unknown): IncidentDetailData | null {
-  if (!incident || typeof incident !== "object") return null;
-  const d = incident as Record<string, unknown>;
+  const match = await prisma.match
+    .findFirst({
+      where: { slug: matchSlug },
+      select: { homeTeam: true, awayTeam: true, incidents: { select: { id: true, slug: true, type: true, minute: true } } },
+    })
+    .catch(() => null);
+
+  if (!match) return { title: "Pozisyon | Var Odası" };
+
+  const shortId = getShortIdFromIncidentSlug(incidentSlug);
+  const inc = match.incidents.find(
+    (i) => i.slug === incidentSlug || (shortId && i.id.endsWith(shortId))
+  );
+
+  if (!inc) return { title: "Pozisyon | Var Odası" };
+
+  const typeLabel = TYPE_LABELS[inc.type] ?? inc.type;
+  const minuteStr = inc.minute != null ? `${inc.minute}. dk ` : "";
+  const title = `${minuteStr}${typeLabel} — ${match.homeTeam} vs ${match.awayTeam} | Var Odası`;
+  const description = `${match.homeTeam} - ${match.awayTeam} maçında yaşanan ${minuteStr}${typeLabel.toLowerCase()} pozisyonunu analiz edin.`;
+  const canonical = `${SITE_URL}/matches/${matchSlug}/${incidentSlug}`;
+
   return {
-    id: String(d.id ?? ""),
-    type: String(d.type ?? ""),
-    minute: typeof d.minute === "number" ? d.minute : null,
-    description: String(d.description ?? ""),
-    confidenceScore: Number(d.confidenceScore ?? 0),
-    sources: Array.isArray(d.sources) ? d.sources.map(String) : [],
-    inFavorOf: d.inFavorOf != null && String(d.inFavorOf).trim() ? String(d.inFavorOf).trim() : null,
-    against: d.against != null && String(d.against).trim() ? String(d.against).trim() : null,
-    videoUrl: d.videoUrl ? String(d.videoUrl) : null,
-    relatedVideos: parseJson<{ url: string; title: string }[]>(d.relatedVideos, []),
-    newsArticles: parseJson<{ title: string; url: string; source: string; author: string }[]>(d.newsArticles, []),
-    refereeComments: parseJson<{ author?: string; text: string; sourceUrl?: string }[]>(d.refereeComments ?? [], []),
-    opinions: Array.isArray(d.opinions) ? d.opinions as IncidentDetailData["opinions"] : [],
-    status: String(d.status ?? "PENDING"),
-    match: d.match && typeof d.match === "object" ? (d.match as IncidentDetailData["match"]) : null,
+    title,
+    description,
+    alternates: { canonical },
+    openGraph: { title, description, url: canonical },
   };
 }
 
-export default function IncidentPage({
+function parseJson<T>(raw: string, fallback: T): T {
+  try { return JSON.parse(raw); } catch { return fallback; }
+}
+
+export default async function IncidentPage({
   params,
 }: {
   params: Promise<{ id: string; incidentSlug: string }>;
 }) {
-  const { id: matchSlug, incidentSlug } = use(params);
-  const router = useRouter();
-  const [incident, setIncident] = useState<IncidentDetailData | null>(null);
-  const [loading, setLoading] = useState(true);
-  const [activeVideo, setActiveVideo] = useState<string | null>(null);
+  const { id: matchSlug, incidentSlug } = await params;
 
-  const fetchData = useCallback(async () => {
-    try {
-      const res = await fetch(
-        `/api/incidents?matchSlug=${encodeURIComponent(matchSlug)}&incidentSlug=${encodeURIComponent(incidentSlug)}`,
-        { cache: "no-store" }
-      );
-      if (!res.ok) {
-        setIncident(null);
-        return;
-      }
-      const data = await res.json();
-      const parsed = toTemplateData(data);
-      setIncident(parsed);
-      if (parsed?.videoUrl) setActiveVideo(parsed.videoUrl);
-    } catch {
-      setIncident(null);
-    } finally {
-      setLoading(false);
+  // Resolve match by slug
+  let match = await prisma.match.findFirst({
+    where: { slug: matchSlug },
+    include: { incidents: { select: { id: true, slug: true } } },
+  }).catch(() => null);
+
+  if (!match) {
+    // Fallback: compute slug from stored fields
+    const allMatchesSlim = await prisma.match.findMany({
+      select: { id: true, slug: true, league: true, week: true, date: true, homeTeam: true, awayTeam: true },
+    }).catch(() => []);
+    const found = allMatchesSlim.find(
+      (m) =>
+        buildMatchSlug({ league: m.league, week: m.week, date: m.date, homeTeam: m.homeTeam, awayTeam: m.awayTeam }) ===
+        matchSlug
+    );
+    if (found) {
+      match = await prisma.match.findUnique({
+        where: { id: found.id },
+        include: { incidents: { select: { id: true, slug: true } } },
+      }).catch(() => null);
     }
-  }, [matchSlug, incidentSlug]);
-
-  useEffect(() => {
-    fetchData();
-  }, [fetchData]);
-
-  useEffect(() => {
-    if (!incident) return;
-    const m = incident.match;
-    const summary = m
-      ? [m.homeTeam, "vs", m.awayTeam, incident.minute != null ? `${incident.minute}. dk` : "", (TYPE_LABELS[incident.type] as { label: string } | undefined)?.label ?? incident.type].filter(Boolean).join(" · ")
-      : [incident.minute != null ? `${incident.minute}. dk` : "", (TYPE_LABELS[incident.type] as { label: string } | undefined)?.label ?? incident.type].filter(Boolean).join(" · ");
-    document.title = summary ? `${summary} | Var Odası` : "Var Odası";
-    return () => {
-      document.title = "Var Odası - Hakem Kararları Analiz Platformu";
-    };
-  }, [incident]);
-
-  if (loading) {
-    return (
-      <div className="flex items-center justify-center py-20">
-        <Loader2 className="h-8 w-8 animate-spin text-red-500" />
-      </div>
-    );
   }
 
-  if (!incident) {
-    return (
-      <div className="mx-auto max-w-4xl px-4 py-20 text-center sm:px-6">
-        <p className="text-lg text-zinc-400">Pozisyon bulunamadı</p>
-      </div>
-    );
-  }
+  if (!match) notFound();
 
-  return (
-    <IncidentDetailTemplate
-      incident={incident}
-      activeVideo={activeVideo}
-      onBack={() => router.push(matchUrl(matchSlug))}
-    >
-      <CommentSection incidentId={incident.id} />
-    </IncidentDetailTemplate>
+  const shortId = getShortIdFromIncidentSlug(incidentSlug);
+  const incidentRow = match.incidents.find(
+    (i) => i.slug === incidentSlug || (shortId && i.id.endsWith(shortId))
   );
+  if (!incidentRow) notFound();
+
+  const dbIncident = await prisma.incident.findUnique({
+    where: { id: incidentRow.id },
+    include: {
+      match: {
+        include: {
+          referee: { select: { id: true, name: true, slug: true, role: true } },
+          varReferee: { select: { id: true, name: true, slug: true, role: true } },
+        },
+      },
+      opinions: { include: { commentator: true }, orderBy: { createdAt: "asc" } },
+    },
+  }).catch(() => null);
+
+  if (!dbIncident) notFound();
+
+  const resolvedVideoUrl = await getResolvedVideoUrl(dbIncident);
+
+  const incident = {
+    id: dbIncident.id,
+    type: dbIncident.type,
+    minute: dbIncident.minute,
+    description: dbIncident.description,
+    confidenceScore: dbIncident.confidenceScore,
+    sources: parseJson<string[]>(dbIncident.sources, []),
+    inFavorOf: dbIncident.inFavorOf ?? null,
+    against: dbIncident.against ?? null,
+    videoUrl: resolvedVideoUrl,
+    relatedVideos: parseJson<{ url: string; title: string }[]>(dbIncident.relatedVideos, []),
+    newsArticles: parseJson<{ title: string; url: string; source: string; author: string }[]>(
+      dbIncident.newsArticles,
+      []
+    ),
+    refereeComments: parseJson<
+      { author?: string; text: string; sourceUrl?: string; stance?: "AGREE" | "DISAGREE" | "NEUTRAL" }[]
+    >(dbIncident.refereeComments, []),
+    opinions: dbIncident.opinions.map((o) => ({
+      id: o.id,
+      comment: o.comment,
+      stance: o.stance,
+      sourceUrl: o.sourceUrl,
+      commentator: {
+        id: o.commentator.id,
+        name: o.commentator.name,
+        slug: o.commentator.slug,
+        role: o.commentator.role,
+      },
+    })),
+    status: dbIncident.status,
+    match: dbIncident.match
+      ? {
+          id: dbIncident.match.id,
+          homeTeam: dbIncident.match.homeTeam,
+          awayTeam: dbIncident.match.awayTeam,
+          league: dbIncident.match.league,
+          week: dbIncident.match.week,
+          date: dbIncident.match.date.toISOString(),
+          referee: dbIncident.match.referee ?? null,
+          varReferee: dbIncident.match.varReferee ?? null,
+        }
+      : null,
+  };
+
+  return <IncidentPageClient incident={incident} matchSlug={matchSlug} />;
 }
