+++ src/app/superlig-tartismali-pozisyonlar-2025/page.tsx
@@ -1,57 +1,53 @@
-"use client";
-
-import { useEffect, useMemo, useState } from "react";
+import { prisma } from "@/database/db";
+import { buildMatchSlug } from "@/lib/slug";
 import Link from "next/link";
-import { Loader2, Flame, Calendar, AlertTriangle } from "lucide-react";
+import { Flame, Calendar, AlertTriangle } from "lucide-react";
 import { matchUrl } from "@/lib/links";
 
-interface MatchSummary {
-  id: string;
-  slug?: string | null;
-  homeTeam: string;
-  awayTeam: string;
-  league: string;
-  week: number;
-  date: string;
-  incidents: { id: string; status: string }[];
-}
+export const revalidate = 300;
 
-export default function SuperLig2025ControversyPage() {
-  const [matches, setMatches] = useState<MatchSummary[]>([]);
-  const [loading, setLoading] = useState(true);
+export default async function SuperLig2025ControversyPage() {
+  const dbMatches = await prisma.match
+    .findMany({
+      where: {
+        league: { contains: "Süper Lig" },
+        date: { gte: new Date("2025-01-01"), lt: new Date("2026-01-01") },
+        incidents: { some: { status: "APPROVED" } },
+      },
+      select: {
+        id: true,
+        slug: true,
+        homeTeam: true,
+        awayTeam: true,
+        league: true,
+        week: true,
+        date: true,
+        incidents: { where: { status: "APPROVED" }, select: { id: true } },
+      },
+      orderBy: { date: "asc" },
+    })
+    .catch(() => []);
 
-  useEffect(() => {
-    async function fetchData() {
-      try {
-        const res = await fetch("/api/matches", { cache: "no-store" });
-        const data = await res.json();
-        const list: MatchSummary[] = Array.isArray(data) ? data : [];
-        const filtered = list.filter((m) => {
-          const year = new Date(m.date).getFullYear();
-          return (
-            m.league.toLowerCase().includes("süper lig") &&
-            year === 2025
-          );
-        });
-        setMatches(filtered);
-      } catch {
-        setMatches([]);
-      } finally {
-        setLoading(false);
-      }
-    }
-    fetchData();
-  }, []);
-
-  const ranked = useMemo(() => {
-    return [...matches]
-      .map((m) => {
-        const approved = m.incidents.filter((i) => i.status === "APPROVED").length;
-        return { match: m, approved };
-      })
-      .filter((x) => x.approved > 0)
-      .sort((a, b) => b.approved - a.approved);
-  }, [matches]);
+  const ranked = dbMatches
+    .map((m) => ({
+      id: m.id,
+      slug:
+        m.slug ??
+        buildMatchSlug({
+          league: m.league,
+          week: m.week,
+          date: m.date,
+          homeTeam: m.homeTeam,
+          awayTeam: m.awayTeam,
+        }),
+      homeTeam: m.homeTeam,
+      awayTeam: m.awayTeam,
+      league: m.league,
+      week: m.week,
+      date: m.date.toISOString(),
+      approved: m.incidents.length,
+    }))
+    .sort((a, b) => b.approved - a.approved);
 
   return (
     <div className="mx-auto max-w-5xl px-4 py-8 sm:px-6">
@@ -69,20 +65,16 @@ export default function SuperLig2025ControversyPage() {
         </p>
       </div>
 
-      {loading ? (
-        <div className="flex items-center justify-center py-16">
-          <Loader2 className="h-8 w-8 animate-spin text-red-500" />
-        </div>
-      ) : ranked.length === 0 ? (
+      {ranked.length === 0 ? (
         <div className="rounded-xl border border-dashed border-zinc-800 py-16 text-center text-sm text-zinc-400">
           Henüz 2025 yılında kayıtlı tartışmalı pozisyon bulunmuyor.
         </div>
       ) : (
         <div className="space-y-4">
-          {ranked.map(({ match, approved }, index) => (
+          {ranked.map(({ id, slug, homeTeam, awayTeam, league, date, approved }, index) => (
             <Link
-              key={match.id}
-              href={matchUrl(match.slug ?? match.id)}
+              key={id}
+              href={matchUrl(slug)}
               className="flex items-center justify-between gap-4 rounded-xl border border-zinc-800 bg-zinc-900/50 p-4 transition-colors hover:border-red-500/40 hover:bg-zinc-900"
             >
               <div className="flex items-center gap-4">
@@ -91,14 +83,14 @@ export default function SuperLig2025ControversyPage() {
                 </div>
                 <div>
                   <p className="text-sm font-semibold text-white">
-                    {match.homeTeam} vs {match.awayTeam}
+                    {homeTeam} vs {awayTeam}
                   </p>
                   <p className="mt-1 flex items-center gap-2 text-xs text-zinc-500">
-                    <span>{match.league}</span>
+                    <span>{league}</span>
                     <span>·</span>
                     <span className="flex items-center gap-1">
                       <Calendar className="h-3.5 w-3.5" />
-                      {new Date(match.date).toLocaleDateString("tr-TR", {
+                      {new Date(date).toLocaleDateString("tr-TR", {
                         day: "numeric",
                         month: "short",
                         year: "numeric",
@@ -120,4 +112,3 @@ export default function SuperLig2025ControversyPage() {
     </div>
   );
 }
-
