+++ src/app/yorumcular/page.tsx
@@ -1,7 +1,21 @@
+import type { Metadata } from "next";
 import { prisma } from "@/database/db";
 import YorumcularClient, { type CommentatorListItem } from "./YorumcularClient";
 
 export const revalidate = 300;
+
+export const metadata: Metadata = {
+  title: "Uzman Yorumcular | Var Odası",
+  description:
+    "Tartışmalı pozisyonları değerlendiren eski hakemler ve spor yorumcularının analizlerini inceleyin.",
+  alternates: { canonical: "https://varodasi.com/yorumcular" },
+  openGraph: {
+    title: "Uzman Yorumcular | Var Odası",
+    description:
+      "Tartışmalı pozisyonları değerlendiren eski hakemler ve spor yorumcularının analizlerini inceleyin.",
+    url: "https://varodasi.com/yorumcular",
+  },
+};
 
 function parseJson<T>(raw: string, fallback: T): T {
   try { return JSON.parse(raw); } catch { return fallback; }
